T-52871: Remove coverage threshold
It would be unproductive insisting on having tests for state
management and components at this point. The presentational
components will be swapped with cs-ui components which would invalidate
any test anyway. Also, the final structure of the application state is
unknown and by locking it down with reducer tests it would make
it much harder to evolve towards an optimal structure as each change
would require updating all of the reducer use-cases.

Change-Id: I2c94e0542cca20a61dd9b93d5ef090a720ef3749